Update hdhr_VCR.applescript
Added "HD" decriptor when viewing channel list
Added more to the setup.  We need to try and get all of the OS prompts cleared using this.
When setting the shows directory, we open up in /Volumes/.  This however is write protected, so we prevented the user form selecting this.
After we start recording, we usd to get a "recording in progress" notification right away.  This is now 2 minutes after the recording starts, and then every 5 minutes.