Update hdhr_VCR.applescript
I am ready to release this. It aint perfect, but handles most conditions pretty well.  We need to flush out the save file, and use jsonhelper to help us here.
I am comfortable enough with #4 being resolved.
Now that we can tuner status, and then distill that down to a handler that returns the next "estimated" time the tuner will be avilable.  We then puase the idle queue, and pad a few seconds, to give us time for the hdhr tuner to become avilable again.  This still isnt 100%, and may be kind of a pain the ass to implment correctly.