20210530
Feature: We will now run a "loop" when we startup, to make sure we start recording shows as quickly as possible.
Feature: Added method to refresh tuners whenever we need to, mainly for testing.
Fix: If we try to record or update a show, and that channel is no longer listed on the lineup, we fail for gracefully.
Fix: Fixed issue where we may set the show_transcode to a null value, rather than "None".

Update: Displaying times in a more reasonable way (rather then saying something is 2400 seconds away)