Not a release, bob
Call the idle handler once before starting the main(0 loop
Made the script more tolerant of issues where there is no tuner on the network, or the tuner is old, or if the tuner returns bad information.  This is, as always, a work in progress, but we took a chunk out of it.
Changed some icons in the dialogs to better reflect the importance of them.
fixed a bug in is_recording
Prevented a crash when a tuner is set to record, that is no longer on the network.
Added more timeouts around network requests