Update hdhr_VCR.applescript
Add: When printing log lines, we will now include if script editor, or the app was running the process
--This allowed us to make it so if script editor is running the script (to debug), we would not actually fire off a curl record job.

Add: You will now see a notification if your HDHR device has a firmware update available.

Fix: Much better handling of not detecting a device on the network, a device that was seen, but is now disappeared, Internet connection being down.

Fix Much better error handling, and  better logging.

Add: We should now detect legacy devices, and not blow up.  This is the result of #26

Soon: is_active true/false for each tuner.  We will then be able to see legacy tuners, but know that they are not compatible.  Not yet implemented yet.