Release 20210905
20210905 is a release!

A few things.
There has been a lot of work to ensure that the script doesn't freak out when a hdhomerun device is not detected on the network.  I have also tried hard to handle situations where the host mac looses a network connection.  Due to the nature of AppleScript, and the overall complexity of the scripts, this is not perfect.  Please report any issues via github issues.

The biggest takeaway since our last official release, which was version 20210805

New:
We now write out config files in a JSON format.  This DRASTICALLY reduces complexity, and allows us to add many more switches to the script, allowing for great improvements in the future.
We will try to read your old config file, and convert to to a JSOn file.  This may fail, and if it does, I am sorry.  In this case, remove the hdhr_VCR.config and hdhr_VCR.json file, and relaunch the script.
When adding shows, we will show the stats about what, why, how, etc of the recording.
Multiple shows adding
Multiple shows editing
Show icons when adding shows
Added more progress bars.

We now rotate the log files, every 5000 + 200 lines per show added.  This isn't perfect, but will hopefully reduce the amount of space this file can take up.