Update hdhr_VCR.applescript
Change: Updated quit handler to kill a show by PID, instead of killing all curl jobs

Change: Changed most of the quotes in the UI to use a variables, instead of hard coded.

New: We now have variables online_detected, and hdhr_online so we can better know our current state.

Change: New icons for things.  I am not in love with that is chosen, so if there are any other unicode icons that are better, let me know.

Fix: I think i finally solved the line endings problem I have had forever, time will tell.  This closes #43

Fix: Trimming of the logs is now easier due to resolution of #43

New: We now version the config, this work is started in #40

New: We now check disk space, and will stop new shows from recording if you are 95% or greater full on the selected folder for the show. (beta)