Update hdhr_VCR.applescript
Remove: Removed Periodic save, and just aligned it with the guide updates on the 15's and 45's
Changed some internal handler functions and names
Add: Working to add SeriesID to recorded shows, if possible.  This would allow us to record a single, a series, or a TV sho(regardless of channel).  This  is still in dev, but wanted to start it.
If we deactivate a show, while it is recording, we will now kill that recording, and mark it not recording