Update hdhr_VCR.applescript
Change: Stops user if they are not in the en_US region
Changed: Moved code for upnext to hopefully run.
New: Added is_sport, so we can track shows that are sports, and add more time to them.
Change: Removed handler that functionality was duplicated
New: Added show_time_orig, so we do not "slide" into another timeslot.
New: Added recording end time to curl request, as a header
Fix: Ensured is_sport is a boolean
NEW: Added code to hopefully allow us to track future shows, so we do not setup too many recordings.  This is inactive for now