Update hdhr_VCR.applescript
Change: Made it more clear in logs, if we had a mismatch on recording shows and tuners.
Fix: Fixed error diaslog when clicking "Run" when selecting a channel to record on.
New: Added a record icon on channel list, if we are currently recording a show.
Change: Removed vuild_channel_list from standard json refresh.  We now run this when the user clicks "Add..."