Update hdhr_VCR.applescript
Change: broke out the run handler into multiple, to aid in allowing a config file reload in the future.
New: If a show is recording, and we attempt to quit the script, we will will show a window, noting we are recording.  This caused an issue when halting a reboot or shutdown.  We can now see if the system is being restarted or shutdown, and skip asking the user.
New: Added idle_uniq and run_uniq to better track loops.
New: Added a way to dump detected tuners into the log
Fix: If a show is recording, and that show has not been added, we now will log a line, and not error out.
New: Added a few new log levels, to match those used with jlogger.
New: Added a way to change the log level via setup()
Fix: Changed =, < and > to words in script.