Update hdhr_VCR.applescript
Add: new item in show json, show_url, which is the full URL as provided by the hdhr device.  Previously, we were building our own URL based on the other information gathered.  This assumed the port of 5004, which is not always true.
Add: Better logging for shows in progress
Add: Updates to channel list, and adding shows.
Removed: Old code.  We are hitting the upward limits of applescript, so only bugs will get fixed now, as adding new code, means removing other code.
New: New handler for "skipping" the delay in the idle queue.  This is helpful when adding shows, but before we pop up the main() window again