Update hdhr_VCR.applescript
Fix: We had to remove more stale code, as we are hitting the script editor script length limit.
Fix: Fix for Sonoma.  The date object, when converted to text, adds a narrow, non breaking space, which was not expected.
Update: Fixed how we determine if the script is running standalone, or in an editor