Update hdhr_VCR.applescript
I had to remove lots of outdated and code comments, as I am quickly hitting the max file size with Script Editor.
Changes:
Changed how we determined if the script was running in a script editor, or debugger, compared to a normal run.
Changed a lot of log lines to TRACE, to reduce the noisiness of DEBUG
Changed when we update the lineup and the guide data.
Changed how we interact with the hdhr api
Reduced the complexity of some dialogs == This may cause issues when using with a non extend device.
Removed support, for now, of adding more time to a sports program