Changes are a comin
Add: Added JSOn logging level.  if this is enabled, all the API calls to https://api.hdhomerun.com and the local device, will be opened in the default web browser.
Add: Added handlername is each handler, as a way of cleaning up the logger statements.
Add: We now use an library file, which works around a text size limit in the script
Add: We will soon be able to record shows, based on the series, on a specific channel, or any channels.  If Mister CSI Miama is on, you can select it, and then choose to record it as a series, but will be asked if you want to record all instances of this show on this channel, or all channels.  This is great for shows that randomly broadcast on different channels, but may not be at a specific day or time.
Add: show_use_seriesid and show_use_seriesid_all to the config, to support the  new seriesid record option.
Add: If a show fails 5 times to record, we will stop trying, to prevent lots of notifcations and extra work